feat: initial release
This library is a port of the prisma-nested-middleware library to work
with Prisma extensions.

The core is largely the same with the args being parsed to determine any
nested operations, however the API for how to modify those nested
operations has changed. Where in prisma-nested-middleware the
createNestedMiddleware function took a middleware function and returned
a new middleware function to pass to $use, this library has a
withNestedOperations function that returns a callback that is passed to
the query.$allModels.$allOperations hook in an extension.

The withNestedOperations function takes an object with two callbacks:
$rootOperation and $allNestedOperations. The nested operations are
passed to $allNestedOperations and then the result of those callbacks
are used to modify the args used in the $rootOperation callback.
The result from the $rootOperation query is similarly modified with the
results from the $allNestedOperations query functions.

To use the types from previous extensions in withNestedExtensions it is
possible to pass ExtArgs as the first type.

In future hopefully it will be possible to create an extension that
modifies $extends directly, however currently that is not possible due
to difficulties in inferring ExtArgs from the $extends this type.