fix: changing operation in $rootOperation
In $rootOperation the only way to change the operation type is to use
the previous client; in $allNestedOperations we are able to manipulate
the args that are passed to the query, but in $rootOperation there is no
API for switching the operation type of the query function.

Currently if the $rootOperation does not call the query function then
the query hangs, this is because we are awaiting the query function in
the $allOperations hook generated by withNestedOperations.

Replace the execution of the $rootOperation with a straight call to
$rootOperation as we don't need to track any aspect of its execution.